Fixed bugs in database functions, simplified encryption functions, and added functionality for storing tokens using Redis

+++ RESTServer.py
@@ -2,22 +2,29 @@ import falcon
 import DatabaseInteractions
 import EncryptionFunctions
 import os
+import secrets
+import redis
 
+if os.path.exists(DatabaseInteractions.database_file):
+    print(str('Database exists'))
+else:
+    print(str('Creating database'))
+    DatabaseInteractions.create_database()
+
+token_db = redis.StrictRedis(host='localhost', port=6379, db=0)
 
 class BaseResource(object):
 
     def on_get(self, req, resp):
         session_token = req.get_param('session_token')
-        # authenticate if possible
-        username = 'username'
-        additional_info = '{phone_number:(123)456-7890}'
-        authenticated = False
-        if authenticated:
-            DatabaseInteractions.retrieve_user(username)
+        if session_token is not None and self.authenticate(session_token):
+            username = token_db.get(session_token)
+            user_data = DatabaseInteractions.retrieve_user(username)
             resp.status = falcon.HTTP_200
-            resp.body = '{username: ' + username + ', additional_info: ' + additional_info + '}'
+            resp.body = '{username: ' + username + ', additional_info: ' + user_data['additional_info'] + '}'
         else:
             resp.status = falcon.HTTP_200
+            resp.content_type = 'application/json'
             resp.body = '{response: "Hello World"}'
 
 
@@ -34,46 +41,64 @@ class UserResource(object):
         DatabaseInteractions.create_user(username, enc_password, salt, additional_info, is_admin)
         resp.status = falcon.HTTP_200
 
-    def on_get(self, req, resp):
-        # authenticate
-        username = 'username'
-        resp.status = falcon.HTTP_200
-        resp.body = DatabaseInteractions.retrieve_user(username)
+    def on_get(self, req, resp, username):
+        token = req.get_param('session_token', True)
+        if self.authenticate(token) and token_db.get(token) == username:
+            user_data = DatabaseInteractions.retrieve_user(username)
+            if not user:
+                raise falcon.exceptions.HTTPNotFound()
+            resp.status = falcon.HTTP_200
+            resp.content_type = 'application/json'
+            resp.body = user_data
+        else:
+            resp.status = falcon.HTTP_401
 
+    def on_put(self, req, resp, username):
+        token = req.get_param('session_token', True)
+        if self.authenticate(token) and token_db.get(token) == username:
+            additional_info = '{additional: info}'
+            DatabaseInteractions.update_user(username, additional_info)
+            resp.status = falcon.HTTP_200
+        else:
+            resp.status = falcon.HTTP_401
 
-    def on_put(self, req, resp):
-        # authenticate
-        for param in req.get_params():
-            if param(0) == 'username':
-                username = param(1)
-            if param(0)=='password':
-                password = param(1)
-
-        username = 'username'
-        additional_info = '{additional: info}'
-        resp.status = falcon.HTTP_200
-        DatabaseInteractions.update_user(username, additional_info)
-
-    def on_delete(self, req, resp):
-        # authenticate
-        username = 'username'
-        DatabaseInteractions.delete_user(username)
-        resp.status = falcon.HTTP_200
-        # if not authenticated, return unauthorized
+    def on_delete(self, req, resp, username):
+        token = req.get_param('session_token', True)
+        if self.authenticate(token) and token_db.get(token) == username:
+            DatabaseInteractions.delete_user(username)
+            resp.status = falcon.HTTP_204
+        else:
+            resp.status = falcon.HTTP_401
 
 
 class AuthenticationResource(object):
 
     def on_post(self, req, resp):
+        username = req.get_param('username', True)
+        password = req.get_param('password', True)
+        retrieved_user = DatabaseInteractions.retrieve_user(username)
+        if retrieved_user is None:
+            resp.status = falcon.HTTP_401
+            return
+        if not EncryptionFunctions.check_password(password, retrieved_user[2], retrieved_user[3]):  # index 2 is password, index 3 is salt for hash
+            resp.status = falcon.HTTP_401
+            return
+        session_token = secrets.token_urlsafe()
+        token_db.setex(session_token, 60, username)  # 24-hour expiration
         resp.status = falcon.HTTP_200
+        resp.content_type = 'application/json'
+        resp.body = '{session_token: ' + session_token + '}'
 
     def on_delete(self, req, resp):
-        # authenticate
-        resp.status = falcon.HTTP_200
+        token = req.get_param('session_token', True)
+        if self.authenticate(token):
+            token_db.delete(token)
+            resp.status = falcon.HTTP_204
+        else:
+            resp.status = falcon.HTTP_401
 
-    def authenticate(self, username, password):
-        # query on username, compare password using encryption
-        return 0
+    def authenticate(self, token):
+        return token_db.get(token) is None
 
 
 app = falcon.API()
@@ -83,10 +108,5 @@ user = UserResource()
 auth = AuthenticationResource()
 
 app.add_route('/', default)
-app.add_route('/user', user)
+app.add_route('/user/{username}', user)
 app.add_route('/auth', auth)
-
-if os.path.exists(DatabaseInteractions.database_file):
-    str("Database exists")
-else:
-    DatabaseInteractions.create_database()
